remove clear after every repeated commands
I removed the clear which got introduced ages ago with a bad amp and/or
serial converter/pc/kernel who knows... Seems this really isn't needed
anymore.

This commit also adds a check to avoid reporting errors if a clear
command was issued with a 'direct' as part of a sequence. Because of the
change of this commit it's unlikely to happen again but one never
knows...

Signed-off-by: Brendan Le Foll <brendan@fridu.org>